requirements.txt
openai
termcolor
python-dotenv
streamlit